Move decorator to util
